+++ src/page/TodayProblem.tsx
@@ -1,5 +1,4 @@
 import { DefaultLayout } from '@layout/index';
-import { useState, useEffect, useCallback } from 'react';
 import {
   TodayProblemHeader,
   TodayProblemNavigationTabs,
@@ -9,38 +8,12 @@ import {
   ChevronLeftIcon,
   ChevronRightIcon,
 } from '@heroicons/react/24/outline';
-import { getTodayProblems } from '@api/problems-v2';
-import useLoadingModal from '@hook/modal/useLoadingModal';
-import { TodayProblem } from '@/type/Problem.type';
+import { useTodayProblem } from '@hook/today-problem/useTodayProblem';
 
 function App() {
-  const [todayProblems, setTodayProblems] = useState<TodayProblem[]>([]);
-  const [currentProblemIndex, setCurrentProblemIndex] = useState(0);
-
-  const currentProblem = todayProblems[currentProblemIndex];
-  const { startLoading, endLoading } = useLoadingModal();
-  const updateTodayProblems = useCallback(async () => {
-    startLoading();
-    try {
-      const response = await getTodayProblems();
-      const todayProblems = response.data;
-      setTodayProblems(todayProblems);
-    } finally {
-      endLoading();
-    }
-  }, []);
-
-  useEffect(() => {
-    updateTodayProblems();
-  }, []);
-
-  const nextProblem = () => {
-    setCurrentProblemIndex((prev) => (prev + 1) % todayProblems.length);
-  };
-
-  const prevProblem = () => {
-    setCurrentProblemIndex((prev) => (prev - 1 + todayProblems.length) % todayProblems.length);
-  };
+  const {
+    todayProblems, currentProblemIndex, nextProblem, prevProblem, setCurrentProblemIndex,
+  } = useTodayProblem();
 
   return (
     <DefaultLayout>
@@ -74,7 +47,7 @@ function App() {
             </button>
 
             <TodayProblemCard
-              problem={currentProblem}
+              problem={todayProblems[currentProblemIndex]}
             />
           </div>
